requirements.txt
Flask==3.1.3
flask-socketio==5.6.1
flask-cors==6.0.2
firebase-admin==7.1.0
openai==2.21.0
python-dotenv==1.2.1
simple-websocket==1.1.0
